Refactor code
Move DownloadLinkerd2 func to pkg/get/linkerd2
